// helper/enum.ts
export enum SupplierComplianceDocuments {
    Profile = "profile",
    Drug = 'drug',
    AlcoholPolicy = 'alcohol-policy',
    Procedure = 'procedure',
    RiskManagementPolicy = 'risk-management-policy',
    SpeedPolicy = 'speed-policy',
    FatiquePolicyPresentationSystem = 'fatique-policy-presentation-system',
    GpsSnapshot = 'gps-snapshot',
    WorkHealthSafetyPolicy = 'work-health-safety-policy',
    RegoDocuments = "rego-documents",
    LicenseDocuments = 'license-documents',
    Documents = 'documents',
    AccreditationDouments = 'accreditation-douments',
    ProductLiability = 'product-liability',
    PublicLiability = 'public-liability',
    WorkCover = 'work-cover',
    Marine = 'marine',
    MarineAlcohol = 'marine-alcohol',
    Coc = 'coc',
    ComplianceDocuments = 'compliance-documents'
}

export enum DriverDocuments {
    Profile = "profile",
    LicenseDocuments = 'license-documents',
    OnoardingDocuments = "onboarding-documents"
}

export enum CustomerDocuments {
    Profile = "profile",
    ContractualDocuments = "contractual-documents",
}

export enum SupplierDriverOnboardingDocuments {
    VisaStatus = "visa-status",
    DriverLicenseFront = "driver-license-front",
    DriverLicenseBack = "driver-license-back",
    LicenseHistory = "lincese-history",
    PoliceVerification = "police-verification",
    PassportFront = "passport-front",
    PassportBack = "passport-back",
    HealthInsurance = "health-insurance",
    DriverCertificate = "driver-certificate",
    Fitness = "fitness",
    DrugTest = "drug-test",
    OnboardingDocuments = 'onboarding-documents'

}

export enum DriverOnboardingDocuments {
    VisaStatus = "visa-status",
    DriverLicenseFront = "license-front",
    DriverLicenseBack = "license-back",
    LicenseHistory = "lincese-history",
    PoliceVerification = "police-verification",
    PassportFront = "passport-front",
    PassportBack = "passport-back",
    HealthInsurance = "health-insurance",
    DriverCertificate = "certificate",
    Fitness = "fitness",
    DrugTest = "drug-test",
}